+++ packages/payer-logo/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.3](https://github.com/Availity/availity-react/compare/@availity/payer-logo@1.1.2...@availity/payer-logo@1.1.3) (2019-02-26)
+
+
+### Bug Fixes
+
+* **payer-logo:** correct path from payerID query ([a6edbaf](https://github.com/Availity/availity-react/commit/a6edbaf))
+
+
+
+
+
+## [1.1.2](https://github.com/Availity/availity-react/compare/@availity/payer-logo@1.1.1...@availity/payer-logo@1.1.2) (2019-02-25)
+
+
+### Bug Fixes
+
+* **payer-logo:** make ajax call to get appropriate logo ([37e154c](https://github.com/Availity/availity-react/commit/37e154c))
+* removed npmrc and added config to lerna for now creating locks ([efe82ab](https://github.com/Availity/availity-react/commit/efe82ab))
+* removed package-locks,added npmrc ignore,fixed react as hoisted dep ([4648319](https://github.com/Availity/availity-react/commit/4648319))
+
+
+
+
+
 ## [1.1.1](https://github.com/Availity/availity-react/compare/@availity/payer-logo@1.1.0...@availity/payer-logo@1.1.1) (2019-02-13)