+++ packages/upload/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.2](https://github.com/Availity/availity-react/compare/@availity/upload@1.7.1...@availity/upload@1.7.2) (2019-02-25)
+
+
+### Bug Fixes
+
+* removed npmrc and added config to lerna for now creating locks ([efe82ab](https://github.com/Availity/availity-react/commit/efe82ab))
+* removed package-locks,added npmrc ignore,fixed react as hoisted dep ([4648319](https://github.com/Availity/availity-react/commit/4648319))
+
+
+
+
+
 ## [1.7.1](https://github.com/Availity/availity-react/compare/@availity/upload@1.7.0...@availity/upload@1.7.1) (2019-02-13)